feat: install groff package after remove groff-base
Previously when we try to install `groff` package, the build step still
showing `groff-base` instead of `groff`. And then, we try to remove the
`groff-base` package and the build step throw an error as expected.

So let's try installing `groff` after removing `groff-base` and see the
result from build step.