docs(readme): guideline for release more focus on minor and patch
Feels like the previous guideline could increase the major version quite
significantly, especially if we add or removing items in the section
quite often. So let's try focus more on minor and patch number for
changes to the section and only increase the major version when we add
or remove section.