feat: add career break in experience and remove pyv in projects
I planning to switch career and that's why currently on career break.
And because we can't generate the pdf file in one page, and i am not
sure how to change the bottom margin yet, so we need to remove some info
for the resume to stay in one page. The info being removed is pyv in
projects section.